finished the project added the update function and sepreated teh logic from the touters
